show all users in users page and restrict page access
