+++ src/features/users/userService.js
@@ -9,7 +9,6 @@ const get = async (param) => {
 
 const getAll = async () => {
     const response = await authenticator.setAuthorization.get(`${API}`)
-    console.log(response)
     if (response.data.success === false) throw new Error(response.data.message)
     return response.data.users
 }
